% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ComputeDistances_1Network_1Tree.R
\name{Function_ComputeDistances_1Network_1Tree}
\alias{Function_ComputeDistances_1Network_1Tree}
\title{Function_ComputeDistances_1Network_1Tree: function to compute probabilistic distances between a network and a bifurcating tree}
\usage{
Function_ComputeDistances_1Network_1Tree(
  list.Model_01_Network,
  list.Model_02_Tree
)
}
\arguments{
\item{list.Model_01_Network}{List containing the following (1) handle.Phylogeny, (2) string.Model = "BM", "OU", "EB", "lambda", "kappa", "delta", (3) vector.Z = vector of mean (ancestral) state values, and (4) vector.Model_01_Theta = vector containing relevant parameters for the models}

\item{list.Model_02_Tree}{List containing the following (1) handle.Phylogeny, (2) string.Model = "BM", "OU", "EB", "lambda", "kappa", "delta", (3) vector.Z = vector of mean (ancestral) state values, and (4) vector.Model_01_Theta = vector containing relevant parameters for the models}
}
\value{
vector.Distances Vector containing the distances computed between the two focal tree models\cr
}
\description{
This function returns a vector containing the Hellinger and KL distances between two tree models
}
\examples{


################
# Load depends #
################
library(ape)
library(geiger)
library(gaussDiff)
library(BMhyd)

######################
# Simulate a network #
######################
handle.SimulatedNetwork <- SimulateNetwork(ntax.nonhybrid=5, ntax.hybrid=3,
                                           flow.proportion=0.5, origin.type='clade', birth = 1, death = 0.5, sample.f = 0.5,
                                           tree.height = 1, allow.ghost=FALSE)

###############################
# Set the first network model #
###############################
list.Model01_BM <- list(handle.Phylogeny = handle.SimulatedNetwork$phy,
                        handle.Flow = handle.SimulatedNetwork$flow,
                        string.Model = "BM",
                        vector.Z = rep(0, 8),
                        numeric.Sig2 = 1)

#############################
# Set the second tree model #
#############################
list.Model02_BM <- list(handle.Phylogeny = handle.SimulatedNetwork$phy,
                        string.Model = "BM",
                        vector.Z = rep(0, 8),
                        numeric.Sig2 = 1)


#####################
# Compute distances #
#####################
Function_ComputeDistances_1Network_1Tree(list.Model_01_Network = list.Model01_BM, list.Model_02_Tree = list.Model02_BM)

}
\keyword{brownian}
\keyword{continuous}
\keyword{distance,}
\keyword{distances,}
\keyword{evolution}
\keyword{model}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{probabilistic}
\keyword{trait}
