% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ComputeDistances_2Networks.R
\name{Function_ComputeDistances_2Networks}
\alias{Function_ComputeDistances_2Networks}
\title{Function_ComputeDistances_2Networks: function to compute probabilistic distances between two network models}
\usage{
Function_ComputeDistances_2Networks(
  list.Model_01_Network,
  list.Model_02_Network
)
}
\arguments{
\item{list.Model_01_Network}{List containing network model with (1) phylogenetic tree (2) table of migratio flow (3) vector.Z and (4) Sig2 parameter}

\item{list.Model_02_Network}{List containing network model with (1) phylogenetic tree (2) table of migratio flow (3) vector.Z and (4) Sig2 parameter}
}
\value{
vector.Distances Vector containing the distances computed between the two focal tree models\cr
}
\description{
This function returns a vector containing the Hellinger and KL distances between two tree models
}
\examples{


################
# Load depends #
################
library(ape)
library(geiger)
library(gaussDiff)
library(BMhyd)

######################################
# Specific list for network model 01 #
######################################
list.Model_01_Network <- list(handle.Phylogeny = handle.SimulatedNetwork$phy,
                              handle.Flow = handle.SimulatedNetwork$phy,
                              vector.Z = rep(0, length(network$phy$tip.label)),
                              numeric.Sig2 = 1)

list.Model_02_Network <- list(handle.Phylogeny = handle.SimulatedNetwork$phy,
                              handle.Flow = handle.SimulatedNetwork$phy,
                              vector.Z = rep(0, length(network$phy$tip.label)),
                              numeric.Sig2 = 2)

#####################
# Compute distances #
#####################
Function_ComputeDistances_2Networks(list.Model_01_Network = list.Model_01_Network,
                                    list.Model_02_Network = list.Model_02_Network)

}
\keyword{brownian}
\keyword{continuous}
\keyword{distance,}
\keyword{distances,}
\keyword{evolution}
\keyword{model}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{probabilistic}
\keyword{trait}
