% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Function_ComputePairwiseDistances_SimulateTips_Aldous.R
\name{Function_ComputePairwiseDistances_SimulateTips_Aldous}
\alias{Function_ComputePairwiseDistances_SimulateTips_Aldous}
\title{Function_ComputePairwiseDistances_SimulateTips_Aldous: function to generate pairwise distance computations}
\usage{
Function_ComputePairwiseDistances_SimulateTips_Aldous(
  numeric.NumberReps,
  numeric.n,
  string.Model,
  vector.Theta
)
}
\arguments{
\item{numeric.NumberReps}{Number of simulation replicates}

\item{numeric.n}{Number of tips}

\item{string.Model}{String for evolutonary model}

\item{numeric.birth}{Birth Rate}
}
\value{
List list with (1) matrix.PairwiseDistances_H, (2) matrix.PairwiseDistances_KL, and (3)vector.PairwiseDistances_H
}
\description{
This function returns a list with (1) matrix.PairwiseDistances_H, (2) matrix.PairwiseDistances_KL, and (3)vector.PairwiseDistances_H
}
\examples{



}
\keyword{brownian}
\keyword{continuous}
\keyword{distance,}
\keyword{distances,}
\keyword{evolution}
\keyword{model}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{probabilistic}
\keyword{trait}
