% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ComputePairWiseModelDistances_BM.R
\name{Function_ComputePairWiseModelDistances_BM}
\alias{Function_ComputePairWiseModelDistances_BM}
\title{Function_ComputePairWiseModelDistances_BM: function to compute pairwise model distances between a set of trees given their model parameter estimates}
\usage{
Function_ComputePairWiseModelDistances_BM(
  phylogeny.TreeSet,
  matrix.ModelParams
)
}
\arguments{
\item{phylogeny.TreeSet}{Set of phylogenetic trees}

\item{matrix.ModelParams}{Matrix of model parameters (BM) for each tree in the given set}
}
\value{
List List of pairwise model distances
}
\description{
This function returns a list of pairwise model distances
}
\examples{

################
# Load depends #
################
library(ape)
library(geiger)
library(gaussDiff)

#########################
# Simulate random trees #
#########################
handle.SimulatedRandomTrees <- rmtree(N = 10, n = 10)
vector.SimulatedDataset <- fastBM(tree = handle.SimulatedRandomTrees[[1]], a = 0, mu = 1)

#######################
# Fit models to trees #
#######################
handle.FittedPhylogeneticModels <- Function_FitCont_TreeSet_BM(phylogeny.TreeSet = handle.SimulatedRandomTrees, vector.InputData = vector.SimulatedDataset)

##############################
# Compute pairwise distances #
##############################
handle.Results <- Function_ComputePairWiseModelDistances_BM(phylogeny.TreeSet = handle.SimulatedRandomTrees, matrix.ModelParams = handle.FittedPhylogeneticModels)

}
\keyword{brownian}
\keyword{continuous}
\keyword{distance,}
\keyword{distances,}
\keyword{evolution}
\keyword{model}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{probabilistic}
\keyword{trait}
