% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ComputePairWiseModelDistances_OU.R
\name{Function_ComputePairWiseModelDistances_OU}
\alias{Function_ComputePairWiseModelDistances_OU}
\title{Function_ComputePairWiseModelDistances_OU: function to pairwise distances between a set of trees and fitteed OU models}
\usage{
Function_ComputePairWiseModelDistances_OU(
  phylogeny.TreeSet,
  matrix.ModelParams
)
}
\arguments{
\item{phylogeny.TreeSet}{Set of phylogenetic trees}

\item{matrix.ModelParams}{Matrix of model parameters (OU) for each tree in the given set}
}
\value{
List List of pairwise model distances
}
\description{
This function returns a list of pairwise distances under the OU models
}
\examples{


################
# Load depends #
################
library(ape)
library(geiger)
library(gaussDiff)

#########################
# Simulate random trees #
#########################
handle.SimulatedRandomTree <- rmtree(N = 1, n = 10)
handle.SimulatedRandomTrees <- list()
class(handle.SimulatedRandomTrees) <- "multiPhylo"
handle.SimulatedRandomTrees[[1]] <- handle.SimulatedRandomTrees[[2]] <- handle.SimulatedRandomTrees[[3]] <- handle.SimulatedRandomTrees[[4]] <- handle.SimulatedRandomTrees[[5]] <- handle.SimulatedRandomTree[[1]]
vector.SimulatedDataset <- fastBM(tree = handle.SimulatedRandomTrees[[1]], a = 0, sig2 = 1, alpha = 2)

#######################
# Fit models to trees #
#######################
handle.FittedPhylogeneticModels <- Function_FitCont_TreeSet_0U(phylogeny.TreeSet = handle.SimulatedRandomTrees, vector.InputData = vector.SimulatedDataset)

##############################
# Compute pairwise distances #
##############################
handle.Results <- Function_ComputePairWiseModelDistances_OU(phylogeny.TreeSet = handle.SimulatedRandomTrees, matrix.ModelParams = handle.FittedPhylogeneticModels)



}
\keyword{brownian}
\keyword{continuous}
\keyword{distance,}
\keyword{distances,}
\keyword{evolution}
\keyword{model}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{probabilistic}
\keyword{trait}
