% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ConductSimExperiment_OU_ModelTest.R
\name{Function_ConductSimExperimentOU_ModelTestBM_OU}
\alias{Function_ConductSimExperimentOU_ModelTestBM_OU}
\title{Function_ConductSimExperimentOU_ModelTestBM_OU: function to compute p-values for likelihood ratio test and compute distances}
\usage{
Function_ConductSimExperimentOU_ModelTestBM_OU(
  list.Model_Simulation_OU,
  numeric.NumberOfReps
)
}
\arguments{
\item{list.Model_Simulation_OU}{List containing components of simulation model}
}
\value{
vector.Distances Vector containing the distances computed between the two focal tree models\cr
}
\description{
This function returns a matrix.SimulationResults for the simulation results
}
\examples{

################
# Load depends #
################
#library(ape)
#library(phytools)
#library(geiger)
#library(vioplot)
#library(gaussDiff)
#library(MASS)

##########################
# Build simulation model #
##########################
phylogeny.RandomTree <- pbtree(b = 1, d = 0, n = 50)
vector.Model_Simulation_Theta <- c(1, 10)
names(vector.Model_Simulation_Theta) <- c("Sig2", "alpha")

list.Model_Simulation_OU <- list(handle.Phylogeny = phylogeny.RandomTree,
                                 string.Model = "OU",
                                 vector.Z = rep(0, length(phylogeny.RandomTree$tip.label)),
                                 vector.Theta = vector.Model_Simulation_Theta)


handle.Sim_Results <- Function_ConductSimExperimentOU_ModelTestBM_OU(list.Model_Simulation_OU = list.Model_Simulation_OU, numeric.NumberOfReps = 5)

###################
# Compare this to #
###################
##########################
# Build simulation model #
##########################
phylogeny.RandomTree <- pbtree(b = 1, d = 0, n = 100)
vector.Model_Simulation_Theta <- c(1, 0.0001)
names(vector.Model_Simulation_Theta) <- c("Sig2", "alpha")

list.Model_Simulation_OU <- list(handle.Phylogeny = phylogeny.RandomTree,
                                 string.Model = "OU",
                                 vector.Z = rep(0, length(phylogeny.RandomTree$tip.label)),
                                 vector.Theta = vector.Model_Simulation_Theta)


handle.Sim_Results <- Function_ConductSimExperimentOU_ModelTestBM_OU(list.Model_Simulation_OU = list.Model_Simulation_OU, numeric.NumberOfReps = 5)
}
\keyword{brownian}
\keyword{continuous}
\keyword{distance,}
\keyword{distances,}
\keyword{evolution}
\keyword{model}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{probabilistic}
\keyword{trait}
