% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_FitCont_TreeSet_OU.R
\name{Function_FitCont_TreeSet_0U}
\alias{Function_FitCont_TreeSet_0U}
\title{Function_FitCont_TreeSet_0U: function ro estimate OU model parameters for each tree in a given set}
\usage{
Function_FitCont_TreeSet_0U(phylogeny.TreeSet, vector.InputData)
}
\arguments{
\item{phylogeny.TreeSet}{List of phylogenetic trees}

\item{vector.InputData}{Vector of continuous trait data}
}
\value{
vector.Distances Vector containing the distances computed between the two focal tree models\cr
}
\description{
This function returns a matrix of fitted model parameters (assuming OU) for a given set of trees
}
\examples{


###############
# Load depends #
################
library(ape)
library(geiger)
library(gaussDiff)

#########################
# Simulate random trees #
#########################
handle.SimulatedRandomTrees <- rmtree(N = 10, n = 10)
vector.SimulatedDataset <- fastBM(tree = handle.SimulatedRandomTrees[[1]], a = 0, mu = 1)

handle.FittedPhylogeneticModels <- Function_FitCont_TreeSet_0U(phylogeny.TreeSet = handle.SimulatedRandomTrees, vector.InputData = vector.SimulatedDataset)
}
\keyword{brownian}
\keyword{continuous}
\keyword{distance,}
\keyword{distances,}
\keyword{evolution}
\keyword{model}
\keyword{motion,}
\keyword{phylogenetic}
\keyword{probabilistic}
\keyword{trait}
